Inimigo Feiticeiro e Vida Player
Fiz o inimigo atirar, o tiro machucar o player e um sistema de Last Hit

+++ Time2_2024.1/Assets/Scripts/Enemybase.cs
@@ -8,7 +8,7 @@ public class Enemybase : MonoBehaviour
     SpriteRenderer sr;
     string enemyName;
     string enemyClass;
-    int health;
+    float health;
     int maxHealth;
     int AttackDamage;
 
@@ -23,7 +23,7 @@ public class Enemybase : MonoBehaviour
         sr.color = Color.magenta;
     }
 
-    public void TakeDamage(int damage) 
+    public void TakeDamage(float damage) 
     {
         health -= damage;
         if(health <= 0)
